Moved Main Menu drawing to its own function
